// lib/api/generated/models/models-update-invoice-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ModelsUpdateInvoiceRequest {
    'due_date'?: string;
    'invoice_status': string;
    'notes'?: string;
    'paid_date'?: string;
}